Factor out a stateless InnerRegExpExec
Previously, the ES spec matched RegExps using a RegExpExec internal
algorithm. This algorithm has a couple properties which don't
completely with its usages, leading to awkward workarounds:

- It modifies lastIndex
- There is no mechanism for passing additional flags

Now that "exec" is not called as a method, but rather RegExpExec
is directly invoked from methods like RegExp.prototype[Symbol.split],
it is possible to refactor RegExpExec to create InnerRegExpExec
which does not have these issues.

This patch does that refactoring and takes advantage of it to remove
the lastIndex "save and restore" from RegExp.prototype[Symbol.search]
and, more importantly, the extra RegExp allocation from
RegExp.prototype[Symbol.split]. This is a normative change, since
it affects when the lastIndex property is read and written.