Normative: Improve TypedArray constructor/Proxy interaction
This patch changes the specification text for constructing TypedArrays
to skip traversing the prototype chain to find the TypedArray-related
internal slots, and instead implement the behavior in the individual
_TypedArray_() constructors. Additionally, methods which used to
allocate TypedArrays as outputs directly will now call the species
constructor, since the previous direct construction also depended on
the prototype chain walk. Fixes #163.