Layering: allow global this and global object to diverge
This fixes a long-standing integration bug between HTML and ES wherein ES assumed that the global this value was always the same as the global object. In web browsers (and indeed in most standalone JS engines I know of) this is not the case.

Fixes https://bugs.ecmascript.org/show_bug.cgi?id=1898.