Layering: straighten out initialization
The previous "ECMAScript Initialization()" abstract operation was intended for use in initializing the entire ECMAScript implementation. (This corresponds to an event loop, in browsers.) However, it was ill-suited for its purpose, as each "implementation" does not need a realm implicitly created, and in fact most implementations will have more than one realm. As such, we leave implementation creation up to the host environment, and merge "ECMAScript Initialization()" with InitializeHostDefinedRealm(). Host environments are now able to call InitializeHostDefinedRealm() to create new realms within their own implementations.

In the future, we hope to introduce a new concept that formalizes "ECMAScript implementation", per e.g. e.g. https://esdiscuss.org/topic/event-loops-in-navigated-away-from-windows and https://esdiscuss.org/topic/holy-mixed-metaphors-batman-was-event-loops-in-navigated-away-from-windows.