Editorial: clarify realm vs. Realm Record
Fixes #252. As discussed therein, a lowercase "realm" is an abstract concept, and a "Realm Record" is a specific type of record representing that realm throughout the algorithms and data structures of the spec.

This also removes uses of the term "code realm" in favor of simply "realm", and replaces "the running execution context's Realm" with "the current Realm Record" in a few places since the latter was specifically defined to mean the former.