Normative: Define a RangeError for stack overflow
Introduce a new abstract operation for the the creation of new execution
contexts. Include a new algorithm step in this operation to issue an
abrupt completion in cases where the execution context stack cannot
support additional frames.

The behavior for this case was previously unspecified. Major runtimes
such as SpiderMonkey and V8 already return an abrupt completion
(although they differ in the completion's value). Standardizing this
behavior allows user code to consistently detect stack overflow
exceptions and respond accordingly.