Editorial: Remove duplicate productions.
Specifically, in 7.1.3.1 "ToNumber Applied to the String Type",
remove the productions for:
    DecimalDigits
    DecimalDigit
    ExponentPart
    ExponentIndicator
    SignedInteger
which duplicate productions (modulo the number of colons)
in 11.8.3 "Numeric Literals".

(Note that 7.1.3.1's Syntax already uses nonterminals defined
in 11.8.3 (e.g. BinaryIntegerLiteral), and already has a
sentence referring readers there for the definitions thereof.)

With this change, each nonterminal should have a single defining
production within the body of the spec (i.e., ignoring Annex B),
which should reduce link-target suprises in the rendered spec.

Fixes #370.