Editorial: Many editorial fixes.
fix typo: "ProxyHander" -> "ProxyHandler"

Tweak a few end-of-step parentheticals.

Insert "is *true*" in a few places.
The "is *true*" was occasionally elided in contexts such as
    If IsFoo(_x_), then ...
(presumably by analogy with programming languages, where
explicit comparison with a boolean value is rarely necessary).
But in the vast majority of cases where that could happen, it doesn't.
So change the elided cases to match the non-elided ones.

Delete the phrase "as the argument"
... from
    Let _obj_ be the *this* value as the argument.
because it doesn't appear to mean anything there.

Delete "the production that is"
... from the phrase
    Let _exprRef_ be the result of evaluating the production that is _expr_
because:
(a) it isn't a production, it's a node in a parse-tree
    (i.e., an *instance* of a production); and
(b) the phrase serves no purpose anyway:
    there are lots of occurrences of "the result of evaluating <variable>".

Delete "be" from "Set <X> to be <Y>."
(It only occurs twice, vs hundreds of times without the "be".)

Change parentheses to angle-quotes
... because it's a List.

Insert apostrophe
("a function's behaviour")

Remove underscores from around non-variable words

Add underscores around variable names.

Insert code unit names.
Specifically "(DIGIT ZERO)" for 0x0030
and "(LINE FEED)" for 0x000A.

Change case on a few words.
Mostly upper-to-lower for "set" and "suspend" when not the first word.
But also lower to upper for "Record {...}"

Add period at end of step.

Change comma to period.

Change period to comma
... mostly at the end of a line preceding substeps in an algorithm.

Delete comma before "and" in two-item "list"

Replace "and" with comma
... when it's a non-final connector of a "list".

Insert "then"
... in an "If <condition>, then <substeps>" algorithm step.

Insert semicolon before "otherwise"
... in "If <condition>, ...; otherwise ..." algorithm steps.

Insert comma where it plays the role of "then"
... in "If <condition>, ..." algorithm steps.

Insert comma before "then"
... in "If <condition>, then ..." algorithm steps.

Insert some spaces
... after colon and comma.

Delete some spaces
... before comma, period, right paren, left paren, left square bracket.

Change |FutureReservedWords| to |FutureReservedWord|s
... because the 'FutureReservedWords' is not a defined nonterminal,
so e.g. it doesn't get a hyperlink in the rendered version.

Adjust indentation of two lines.

Add backticks to a few terminals
(referenced in algorithms)