Editorial: Change "[> match only if ..." to "[> but only if ..."
... in the productions for {Lead,Trail,Non}Surrogate
in the regexp grammar.

(There doesn't appear to be a difference in meaning between
the two phrasings, so it's simpler to use just one. And
"match" suggests parsing, which interferes with the generative
interpretation of the grammar, whereas "but" is agnostic on
generation vs parsing. So change "match" to "but".)