[normative] Prevent module binding "rescuing"
Currently, when an indirect export binding cannot be resolved,
ResolveExport continues to search for a binding of the same name as
provided by a "star export." If found, the algorithm exits normally and
that binding is used.

This means that "star exports" can effectively "rescue" faulty indirect
exports, as in the following example:

```js
// file: rescued.js
export { x } from './empty.js';
export * from './has-x.js';
```

```js
// file: empty.js
;
```

```js
// file: has-x.js
export var x;
```

`rescued.js` is ostensibly providing an `x` binding from `empty.js`, but
the binding actually originates in `has-x.js`. The author of this code
likely expects the binding to be provided by `empty.js`, and the fact
that the code executes without error may hide bugs resulting from a
misunderstanding of the actual state of the dependency graph.

Ensure that an error is reported when indirect exports cannot be
resolved, regardless of any "star exports" present in the module.