[editorial] Document invariant more consistently
As noted in the non-normative text, the second argument to
GetBindingValue for module Environment Records will always be true.
However, two other aspects of the text suggest that this is *not* an
invariant:

> [...] a *ReferenceError* is thrown regardless of the value of S.

And:

> 1. Return ? _targetER_.GetBindingValue(_N2_, _S_).

In this latter case, the reference to the variable instead of the
literal value *true* suggests that some other value may be specified.

Replace the non-normative text with a formal assertion as the first step
of the algorithm, and remove the details that tend to contradict the
invariant.