[editorial] Remove unreachable method
The DeleteBinding method of any Environment Record is only called when the
`delete` operator is applied to a value of type Reference. Because module
environment records (and therefore module bindings) are only accessible from
module code, the module Environment Record's version of this method would only
be called if the `delete` expression occurred within module code.

All module code is strict mode code, so the `delete` expression would
necessarily have to exist within strict mode. However, static semantics of the
operator explicitly prohibit its application to an IdentifierReference from
within strict mode.

Because the preconditions for calling this method cannot be satisfied, it may
be removed from the specification.