[editorial] Specify order of module namespace keys
A module Namespace exotic object's [[Exports]] internal slot is defined
through the following sequence of steps:

1. GetModuleNamespace invokes GetExportedNames
2. GetExportedNames returns a List of strings describing the
   LocalExportEntries in source code order, followed by the
   IndirectExportEntries in source code order, followed by the same for
   each of the module's "star" exports.
3. GetModuleNamespace resolves all `exportedNames`, filters out all
   ambiguous names, and passes the result to ModuleNamespaceCreate
4. ModuleNamespaceCreate sets the object's [[Exports]] internal slot to
   the provided list

There is no normative text that modifies the order of list elements
following its creation in GetExportedNames. This conflicts with the
informative text in table 29, "Internal Slots of Module Namespace Exotic
Objects," which describes the [[Exports]] slot as:

> A List containing the String values of the exported names exposed as
> own properties of this object. The list is ordered as if an Array of
> those String values had been sorted using `Array.prototype.sort` using
> SortCompare as _comparefn_.

Because ordering of this list is observable through the
[[OwnPropertyKeys]] method and the ListIterator created by its
[@@iterator] method, the specification should include explicit algorithm
steps that produce this result.

Extend the ModuleNamespaceCreate algorithm to sort the values of
[[Exports]] according to previously-existing text in Table 29.