Editorial: Assert normal completion value  (#507)
Immediately prior to invoking ResolveExport, the [[Get]] method ensures

that the property key exists in the [[Exports]] list. In

ModuleNamespaceCreate, the [[Exports]] slot is set to the list of all

bindings that may be resolved unambiguously from the requested module.

This means that this invocation of ResolveExport cannot return an abrupt

completion.



Document this fact in the text of the algorithm.