[normative] Promote consistent scoping for classes
Prior to evaluating the optional class heritage, the running execution
context's LexicalEnvironment is set to `classScope`, an environment
dedicated to the class's "name" binding. This may be observed from
ECMAScript code such as:

    let C = null;
    {
      (class C extends C {}); // ReferenceError: C is not initialized
    }

The following access of the `prototype` property of the class's parent
constructor (where defined) is an observable operation. As such, the
LexicalEnvironment at this time can be detected here, as well:

    let C = null;
    {
      (class C extends new Proxy(function() {}, { get() { return C; } }) {});
    }

As written, the running execution context's LexicalEnvironment is *not*
set to the environment created for the class's "name" binding (e.g. the
`classScope`). Therefore, this second example would *not* produce a
ReferenceError.

This interleaving lexical environments during ClassDefinitionEvaluation
is inconsistent. Modify the specification text to explicitly re-set the
running execution's LexicalEnvironment when retrieving the `prototype`
property of the parent constructor.