TypedArrays slice get and set operations does not throw
The step 9 from 22.2.3.24 creates a new TypedArray instance that will
be validated on 22.2.4.6 step 2, including a validation for a length
that is >= the given count, which is never gonna be higher than the
original instance length.

It's currently impossible to have an abrupt completion from the get
and set operations on steps 14.b.ii and 14.b.iii.