Layering: take care of edge cases for GetActiveScriptOrModule()
Instead of only looking at Function components, and falling back to the running execution context, simply go for the topmost execution context with a ScriptOrModule component.

The "no such context exists" case can in fact be reached, if GetActiveScriptOrModule() is called with a stack that only contains the "realm execution context" (i.e. if this is called without being initiated by ScriptEvaluation or ModuleEvaluation).