Editorial: spec refactor of PromiseReaction records/PromiseReactionJob
The current spec creates fictional “equivalent functions” called “Identity” and “Thrower”, and passes these around with `PromiseReaction` records.

Instead, this change adds a string “Type” field to `PromiseReaction` records, and sets the “Handler” field to `undefined` when none is provided, so that the behavior can be determined solely inside `PromiseReactionJob`.

This change should be non-observable, and helps future Promise proposals.