Normative: Resolve template argument references
Previously, when performing ArgumentListEvaluation for tagged templates,
the result of evaluating each substitution expression was not passed to
GetValue. This meant that unresolvable references would not interrupt
ArgumentListEvaluation, and that Reference values would eventually be
assigned to function parameter bindings in
FunctionDeclarationInstantiation.

Update ArgumentListEvaluation (and the related SubstitutionEvaluation)
to avoid this situation and increase parity with the abstract
operation's behavior when applied to the Arguments production.