Editorial: Remove impossible condition
Because an anonymous FunctionDeclaration can only occur as part of an
`export default` declaration, an `export` declaration can only occur in
in module code, and module code is always strict mode code, the function
code of an anonymous FunctionDeclaration is always strict mode code.

Remove the condition that allows for non-strict anonymous function
declarations, and update the accompanying informative note to explain
this detail.