Normative: Handle abrupt cmpltn in SuperProperty
During evaluation of SuperProperty, the result of evaluating Expression
may be an abrupt completion. GetValue is designed to immediately return
such completion values, through it may also return a new abrupt
completion if a valid Reference value is provided but cannot be
resolved. In either case, the result of GetValue will be an abrupt
completion, which ToPropertyKey is not designed to handle.

Update the algorithm to immediately return abrupt completions produced
by the invocation of GetValue.

Closes #615.