Remove unused [[ImportEntries]] field of Source Text Module Record.
The top-level ImportEntries of a module are now extracted as part of the
CreateLexicalBindings operation, based on the top-level
LexicallyScopedDeclarations of the module, which now include
|ImportDeclaration| productions in addition to other lexically scoped
declarations like function/generator declarations and let/const variable
declarations.

Though it is possible that another specification might depend on this
field, the only relevant specification that I'm aware of is the WHATWG
Loader spec (https://whatwg.github.io/loader), and while it refers to the
Source Text Module Record in many places, it never makes use of the
[[ImportEntries]] field.

Moreover, I think we should remove this field now rather than allowing it
to persist, because any hypothetical auxiliary specification that relies
on [[ImportEntries]] is likely to be treating top-level import
declarations differently from nested import declarations, and thus runs
the risk of mishandling nested import declarations.