Normative: Clarify rules around super inside eval (#685)
* Normative: Clarify rules around super inside eval

Previously, `super` was allowed inside eval if eval was called inside
any function and disallowed inside arrow functions. This semantics
likely predates the eviction of super from functions during ES6 and was
simply not updated. This PR updates the relevant semantics to align more
closely with both user expectations and implementation reality.

With this PR, SuperProperty is only allowed inside eval if the current
this environment is a function environment record with a home object.
This means SuperCall is allowed in eval (and nestings thereof) in arrow
functions in methods per implementation reality. Other usages of
SuperProperty inside eval is an early error.

Additionally, SuperCall is only allowed inside eval if the current this
environment is a function environment record with a home object and a
[[Construct]] field. Since methods do not have [[Construct]] slots, this
has the effect of only allowing SuperCall inside eval inside
constructors (or inside another eval in a constructor, or inside an
arrow function in a constructor) per implementation reality. Other
usages of SuperCall inside eval is an early error.

Lastly, neither SuperProperty or SuperCall are allowed inside eval
inside a non-method function. This is an early error. This aligns with
implementation reality with the exception of Chakra.

* Editorial: early error conditions -> early error rules