Remove @@iterator from Module Namespace objects
Per consensus at the November 2016 meeting, Module Namespace exotic
objects no longer have an @@iterator method.

Also changes the [[Configurable]] property of the @@toStringTag
property to false, as per consensus in the same meeting (and
because it's impossible to reconfigure due to the exotic
methods of such objects).

This closes #693 and #710.