Normative: do not call super constructor when ThisBindingStatus is already initialized
Given:

```js
class A { constructor() { print('a') } };
class B extends A {
  constructor() {
    super();
    super();
  }
}
```

The current spec says 'a' is printed twice as the second super will
invoke the super constructor before throwing an error due to `this`
being initialized.

This change throws just before invoking the super constructor if this
has already been initialized. The argument list is evaluated as it
may initialize `this` by invoking super (i.e. `super(super())`).

Current implementations agree on the current spec behavior so it's
questionable whether this change is worth doing. However this came up in
real world code (due to a merge mishap) and caused significant confusion
so if we can get away with changing it I think we should.