Editorial: put Memory Model notes in their own subclause
The emu-clause for "Memory Model" contains some introductory stuff,
followed by a bunch of sub-emu-clauses,
followed by 3 emu-notes.
Now, although these notes are (in the HTML)
children of the Memory Model clause,
I suspect that when rendered,
they'll *look* like they're part of
the last of the sub-clauses ("Data Race Freedom"),
which you probably don't want.

So, to prevent this, I put the three notes in their own sub-clause.