Editorial: disambiguate a dangling else
This is a case of "dangling else" (or "dangling otherwise"),
and it's not obvious which "if" the "otherwise" associates with.

I've rewritten it to remove the ambiguity.
Please check whether I got the intent right.
(If not, you might want to rewrite as multiple steps,
so that the logic is clear.)