Editorial: delete "definition" that's just a cross-ref
... for PropertyDefinitionEvaluation.

For one thing, referring to the whole "Generator Function Definitions"
clause is rather vague. Surely only its "RS: PropertyDefinitionEvaluation"
sub-clause is pertinent.

But more to the point, why is the combination of this particular operation
and this particular chain production the only one that warrants a cross-ref?
It doesn't seem that special.