Editorial: "[[AgentOrders]]" -> "[[EventLists]]"
There is no [[AgentOrders]] field.
It's presumably a typo for [[AgentOrder]],
except that's a Relation, which doesn't make sense in this context.
It looks like [[EventLists]] is the only field that does make sense.

(I considered creating an abstract operation for:
    the [[EventList]] field of the element in _execution_.[[EventLists]]
    whose [[AgentSignifier]] is AgentSignifier()
but I'm not sure what it would be called.)