Editorial: separate clauses for Reference abstract ops
Formerly, the 7 low-level abstract operations for References
were defined in a list of one-liners (that used some unusual syntax).
Explode that list, so that each op gets its own emu-clause,
with a fairly conventional algorithm.

For each one, take the 'id' attribute from the former <dfn> element
(to which there might be external references)
and put it on the new <h1> element.