Normative: Relax restrictions on `with` statement
Recent industry trends have shown an increased interest in the use of
the `with` statement as a means for more concise code. Because class
bodies and module code are always strict mode code, this restriction
severely limits usage of this statement.

Remove the strict mode restriction so that the `with` statement may be
leveraged in more contexts.