Normative: Define particular precision for toFixed and friends (#857)
* Normative: Define particular precision for toFixed and friends
Previously, the specification for Number.prototype.toFixed, Number
gave a fixed range for the fractionDigits parameter, but allowed
implementations to permit a larger range. Historically, only
SpiderMonkey accepted a larger range, and other implementations
threw a RangeError outside the range.

This patch specifies the SpiderMonkey behavior. The range checks
always permitted more decimal digits to be printed than what is
provided by the accuracy, so the limits are arbitrary. Waldemar
Horwat has mentioned that SpiderMonkey's higher limits have
been useful for him in explaining floating point numbers to
others.

An alternative would be to retain the previous limits and remove
the implementation-dependent aspects. Either way, locking down
this detail should help developers get consistent behavior
between implementations.

* Switch to 0 as the lower bound for toFixed
As @claudepache points out, the behavior is inconsistent for negative
numbers because the exponential fallback assumes all the digits will
be displayed. The negative support was motivated just by consistency,
without any particular concrete use cases, and it's not supported in
three browsers, so this patch removes it.