Normative: Specify RegExp malformed character class behavior (#864)
This patch specifies a previously unspecified piece of RegExp
syntax and semantics: If a character class has a control escape
(\c) which is not followed by an extended control character
[a-ZA-Z0-9_], then four engines will accept it as syntactically
valid. V8, JSC and SpiderMonkey will treat it as the character
class [\\c]; ChakraCore's behavior differs and depends on what
follows.

This patch codifies the majority behavior within the Annex B
RegExp extensions. This change is consistent with malformed \c
behavior when it takes place outside of a character class: It
behaves like \\c, rather than an identity escape.

Closes #863

* New definition recommended by @anba

* Revert "Normative: Specify RegExp malformed character class behavior"

In a follow-on patch, the new version is given

* Add a lookahead to avoid an ambiguity for [\]