Normative: Clarify that case conversion is context-sensitve
Previously, String.prototype.toUpperCase and
String.prototype.toLowerCase were specified with wording indicating
that Unicode case conversion was to be used, but code points should
be looked up one by one. The Unicode Default Case Conversion algorithm
is context-dependent, and a note alludes to this, but the normative
text seems to prohibit using the context. However, JSC, V8 and
ChakraCore do seem to pay attention to this context, at least when
it comes to final sigma.

This patch clarifies the spec text to make an explicit reference
to the Default Case Conversion algorithm.