Editorial: Refer to proleptic Gregorian calendar.
The spec currently refers to an "extrapolated Gregorian system",
but the term "proleptic Gregorian calendar" is more precise and is
the terminology used elsewhere in other standards and languages.
(e.g. ISO 8601:2004 3.2.1 "The use of this calendar for dates
preceding the introduction of the Gregorian calendar (also called
the proleptic Gregorian calendar) should only be by agreement of
the partners in information exchange.")