Editorial: Add '(API)' after 'application programming interface'.
ECMA 262 Section 2: writes out "application programming interface" without also
writing "(API)" which is relevant because "API" is used throughout the rest of
the document and "(API)" is never made explicit before it's use.

API is a common acronym, so we could rewrite the long-form here to simply "API"
or add "(API)" since this is the first instance (which would be my preference).

Note that this section explicitly references ECMA 402 in the references below
that section, the title of which document is written in ECMA 262 as "ECMA-402,
ECMAScript 2015 Internationalization API Specification".