Meta: Expand CONTRIBUTING.md (#1031)
This patch documents how to contribute to TC39 in a bit more detail,
explaining in terms of how things work today. It does not overlap
with the process document, but instead describes how things work
surrounding that, and provides practical suggestions for getting
involved. Phrasing is changed to include not just non-members but
also Ecma members, who may need this documentation as well. This
document also does not describe how to make a proposal repository,
which will be left to be documented elsewhere (possibly eventually
cross-referenced from here).