Editorial: tweak UnicodePropertyValueExpression eval algs
_UnicodePropertyName_, _UnicodePropertyValue_, and _LoneUnicodePropertyNameOrValue_
are invalid because UnicodePropertyName etc are nonterminals,
not metavariables.

And simply changing them to |UnicodePropertyName| etc
wouldn't be valid, because that would be passing a Parse Node
to UnicodeMatchProperty/UnicodeMatchPropertyValue,
which isn't what they're expecting.

So instead, use the SourceText operation to 'extract'
the List of code points for each.