Editorial: change "a Parse Node" back to "it".
The sentence has two clauses separated by a semicolon.
With the word "it", the second clause has two references
("it" and "that symbol") to the two nouns in the first clause
("Each Parse Node" and "a symbol in the grammar").
If you change "it" to "a Parse Node", it's no longer a reference
to "Each Parse Node", so the second clause becomes confused about
whether or not it's referring to things in the first clause.