Editorial: rework the "unique Parse Node which it covers" sentence.
You can't say "a Parse Node which is an instance of _P_":
_P_ *is* a Parse Node.

Also, as a separate sentence, you might have to say "at most one"
rather than "one".

In trying to rework the sentence, I decided it was easier to drop it,
and instead insert "unique for a given _P_" into the previous sentence.