Editorial: LocalTZA input specification (#1160)
The `LocalTZA` implementation-defined algorithm must accept the full
range of `Number` and not define its input to be a *time value* because
the algorithm is specified to take an already offset local time that can
be outside of the time value range.

This cleans up the specification text to clearly distinguish UTC time
values versus local time `Number` inputs and uses the same subscript
notation for both UTC and local time to make the text more clear.