Normative: Cleanup Time Values and Time Range (#1144)
Implementations are not currently consistently limiting *time values* to
the spec defined ranges of both *time value* and *extended years*.

This change is editorial cleanup that more clearly states the allowed
*time value* range and the supported range of *extended years* that fit
within the allowed *time value* range.

Implementations will need to be updated to conform with the spec (SM
currently adheres) and pass the accompanying test in tc39/test262#1501.

Time Values and Time Range:

- Language about `Number` range in hundreds-of-thousands of years plus/minus
UTC epoch is not immediately relevant; move it to an `emu-note`
- New `emu-note` text fixes inconsistency in spec where this section was
not properly accounting for leap years when computing the approximate number
of years in the maximum range but 20.3.1.15.1 *extended years* was
stating it correctly
- Add `emu-xref` for Number integer range
- Add comment referring to POSIX day for the source of 86,400 seconds
- Simplify remaining spec in section into clear, consistent language

Extended Years:

- Add `dfn` to *extended years* so it is linked from format section
- Add `emu-xref` for *time value* range
- Explicitly state *extended years* relation to *time value* range
- Explicitly state ISO 8601 strings outside *time value* range are not
accepted
- Modify *extended years* example table to reflect the proper
minimum/maximum values allowed by the *time value* range
- More forceful Extended Years text: explicitly state *NaN* is returned
instead of "are not accepted."